modified:   .vscode/settings.json 	new file:   data/background.png 	new file:   data/wifi.png 	new file:   data/wifimanager.css 	new file:   data/wifimanager.html 	modified:   node-red/BME MQTT nodered.json 	modified:   platformio.ini 	modified:   src/00_globals.hpp 	renamed:    src/02_JSON.hpp -> src/01_JSON.hpp 	deleted:    src/01_reboot.hpp 	renamed:    src/03_FileSys.hpp -> src/02_fileSys.hpp 	renamed:    src/04_wifi.hpp -> src/03_wifi.hpp 	renamed:    src/05_bme.hpp -> src/04_bme.hpp 	new file:   src/05_mqtt.hpp 	renamed:    src/07_events.hpp -> src/06_events.hpp 	deleted:    src/06_mqtt.hpp 	modified:   src/main.cpp

+++ src/02_fileSys.hpp
@@ -55,7 +55,7 @@ void fileToCharPtr(fs::FS &fs, const char* path, char* buffer) {
     return;
   }
   if (Debug) Serial.println(F("File"));
-  size_t i = 0;
+  byte i = 0;  // Buffer index (needs to be as big as paramSize (25) -> byte is enough)
   while (file.available() && i < paramSize - 1) {
     buffer[i++] = (char)file.read();
   }
